requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.1
django-extensions==3.2.3
djangorestframework==3.14.0
executing==2.0.1
idna==3.6
ipdb==0.13.13
ipython==8.20.0
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.3
pandas==2.1.4
parso==0.8.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
